requirements.txt
pypdf
groq
pinecone
python-dotenv
sentence-transformers
scikit-learn
streamlit>=1.28.0
numpy
transformers
